// src/lib/api/generated/model/environmentSummarySchema.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The canonical backend API for COMPASS clients.
 * OpenAPI spec version: 1.0.0
 */
import type { EnvironmentComponentsSchema } from './environmentComponentsSchema';
import type { EnvironmentMaintenanceSchema } from './environmentMaintenanceSchema';
import type { ReleaseIdentitySchema } from './releaseIdentitySchema';

export interface EnvironmentSummarySchema {
  components: EnvironmentComponentsSchema;
  deployment_class: string;
  environment: string;
  maintenance: EnvironmentMaintenanceSchema;
  release: ReleaseIdentitySchema;
}