// src/lib/api/generated/model/operationalCommandCatalogSchema.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The canonical backend API for COMPASS clients.
 * OpenAPI spec version: 1.0.0
 */
import type { OperationalCommandSchema } from './operationalCommandSchema';

export interface OperationalCommandCatalogSchema {
  items: OperationalCommandSchema[];
}